Build with Qt6
#8

+++ qjsonmodel.h
@@ -327,7 +327,7 @@ public:
     static QJsonTreeItem* loadByDesc(const QJsonValue& description,
                                      const QStringList &exceptions = {}, QJsonTreeItem * parent = nullptr);
     static JsonFieldType typeFromString(const QString &str);
-    static QVariant defaultFromString(const QString &str, int size);
+    static QVariant defaultFromString(const QString &str, size_t size);
 
 protected:
